typecast 2byte value to 1 byte value
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@4419 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/net/Ping.cxx
@@ -349,7 +349,7 @@ char*			PingPacket::unpackHex8(char* buf, uint8_t& v)
   uint16_t d = 0;
   d = (d << 4) | hex2bin(*buf++);
   d = (d << 4) | hex2bin(*buf++);
-  v = d;
+  v = (uint8_t)d;
   return buf;
 }
